Refactored CDT to use the eclipse-base and updated its artifact name as discussed in #226
