+++ _ext/eclipse-cdt/gradle.properties
@@ -1,16 +1,16 @@
-# Mayor/Minor versions are in line with the one of the Eclipse C/C++ IDE version.
-# Patch version is an incremental counter for the Spotless plugin.
-cdt_version=9.4.0-SNAPSHOT
-cdt_artifactId=spotless-ext-eclipse-cdt
-cdt_description=Eclipse's CDT C/C++ formatter bundled for Spotless
-VER_JAVA=1.8
+# Mayor version correspond to the Eclipse C/C++ IDE version.
+# Minor version changes in case the minimum Eclipse dependencies are raised due to incompatibilities of their internal interfaces.
+# Patch version is incremented in case of bug-fixes and evolutions.
+ext_version=9.0.0-SNAPSHOT
+ext_artifactId=spotless-eclipse-cdt
+ext_description=Eclipse's CDT C/C++ formatter bundled for Spotless
+ext_org=diffplug
+ext_group=com.diffplug.spotless
 
-cdt_org=diffplug
-cdt_group=com.diffplug.spotless
+# Build requirements
+ext_VER_JAVA=1.8
 
 # Compile dependencies
-cdt_VER_ECLIPSE=4.7.2
-cdt_VER_ECLIPSE_CDT=9.4
-
-# Test
-cdt_VER_JUNIT=4.12
+VER_ECLIPSE_CDT=9.4
+VER_SPOTLESS_ECLISPE_BASE=3.+
+VER_ECLISPE_JFACE=3.12.+
